refactor: Cleared up stage namings
