+++ src/lib/modules/loadPackageJSON.ts
