--- src/lib/run/index.ts
@@ -1,4 +1,0 @@
-import runRestartable from './runRestartable';
-import runOnce from './runOnce';
-
-export { runOnce, runRestartable };
